requirements.txt
boto3
npyscreen
yml
pyyaml
pandas
hurry.filesize
click
xlsxwriter